+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 12.8
 - Fix PHP8.5 deprecated using null as an array offset, use an empty string instead in MenuReports.php & ActivityReport.php
 - Fix PHP warning undefined array key when switching Meals in ServeMenus.php
 - Fix SQL error when no Quarters found (Course widget) in Course.php
+- Fix PHP8.5 deprecated trim() Passing null to parameter 1 of type string in StudentFieldBreakdown.php & StudentBreakdown.php
 
 Changes in 12.7.2
 -----------------